requirements.txt
pytest-bdd==7.3.0
requests==2.31.0
selenium==4.14.0
allure-pytest-bdd==2.13.2
